Rename project and namespace for memory pipeline
