+++ memorypipeline/CopilotChatMemoryPipeline.csproj
@@ -1,11 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>disable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>SemanticMemory.Service</RootNamespace>
+    <RootNamespace>CopilotChat.MemoryPipeline</RootNamespace>
     <UserSecretsId>ef47f200-b235-45d9-9fd6-765fc59d33f5</UserSecretsId>
   </PropertyGroup>
