--- memorypipeline/nuget.config
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  
-  <packageSources>
-    <clear />
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <add key="SemanticMemoryPrivate" value="https://pkgs.dev.azure.com/msctoproj/Lightspeed/_packaging/SemanticMemoryPrivate/nuget/v3/index.json" />
-  </packageSources>
-
-  <packageSourceMapping>
-    <packageSource key="nuget.org">
-      <package pattern="*" />
-    </packageSource>
-    <packageSource key="SemanticMemoryPrivate">
-      <package pattern="Microsoft.SemanticMemory.*" />
-    </packageSource>
-  </packageSourceMapping>
-  
-</configuration>
